Disable the Iron Banner command
Disable the iron banner command since bungie have started keeping it up
for two weeks at a time and the bot cannot currently deal with this. The
issue is if the bot assumes a two week period starting on odd weeks and
ending on even weeks, then if Bungie starts an iron banner on an even
week, she will only keep the command up for 1 week.

The alternative that will be used is a manual (User) command definition